# Conflicts:
#	src/chart/generator/plot.cpp
#	src/chart/options/advancedoptions.h
#	src/chart/options/autoparam.h
#	src/chart/options/config.cpp
#	src/chart/options/config.h
#	src/chart/options/optionssetter.h
#	test/qtest/chart.cpp
#	test/unit/base/refl/auto_struct.cpp
#	test/unit/base/style/paramregistry.cpp

+++ src/apps/qutils/canvas.cpp
@@ -197,7 +197,7 @@ void BaseCanvas::setTextColor(const Gfx::Color &color)
 
 void BaseCanvas::beginDropShadow() {}
 
-void BaseCanvas::setDropShadowBlur(uint64_t) {}
+void BaseCanvas::setDropShadowBlur(double) {}
 
 void BaseCanvas::setDropShadowColor(const Gfx::Color &) {}
