+++ src/base/anim/easinggradient.cpp
@@ -26,7 +26,8 @@ EasingGradient EasingGradient::Bezier(const Geom::Point &p1,
 	    Geom::Point(1, 1));
 
 	for (auto i = 0U; i < stepCount; i++) {
-		auto point = bezier(static_cast<double>(i) / (stepCount - 1));
+		auto point = bezier(static_cast<double>(i)
+		                    / static_cast<double>(stepCount - 1));
 		res.stops.emplace_back(point.x, point.y);
 	}
