+++ src/data/datacube/datastat.cpp
@@ -10,10 +10,11 @@ DataStat::DataStat(const DataTable &table,
 	const auto &indices = options.getDimensions();
 	for (const auto &idx : indices) {
 		if (idx.getType().isReal()) {
-			auto valueCnt =
-			    table.getInfo(idx.getColIndex().value()).dimensionValueCnt();
+			auto valueCnt = table.getInfo(idx.getColIndex().value())
+			                    .dimensionValueCnt();
 			usedColumnIDs.insert(
-			    {static_cast<size_t>(idx.getColIndex().value()), usedValues.size()});
+			    {static_cast<size_t>(idx.getColIndex().value()),
+			        usedValues.size()});
 			usedValues.emplace_back();
 			usedValues.back().resize(valueCnt);
 		}
@@ -42,7 +43,8 @@ void DataStat::trackIndex(const TableRow<double> &row,
 	for (auto i = 0U; i < indices.size(); i++) {
 		const auto &idx = indices[i];
 		if (idx.getType().isReal())
-			usedValues[i][row[idx.getColIndex().value()]] = true;
+			usedValues[i][static_cast<size_t>(
+			    row[idx.getColIndex().value()])] = true;
 	}
 }
